fix: fixing up type exports
